+++ deno.jsonc
@@ -1,8 +1,8 @@
 {
   "tasks": {
-    "test": "deno test -A --no-check=remote && deno check --config=www/deno.json www/main.ts && deno check cli.ts",
-    "fixture": "deno run -A --watch ./tests/fixture/main.ts & deno run -A ./cli.ts manifest --watch ./tests/fixture",
-    "www": "deno run -A --watch=www/static,docs/ ./www/main.ts & deno run -A ./cli.ts manifest --watch ./www"
+    "test": "deno test -A --no-check=remote && deno check --config=www/deno.json www/main.ts www/dev.ts && deno check init.ts",
+    "fixture": "deno run -A --watch=static/,routes/,islands/ tests/fixture/dev.ts",
+    "www": "deno run -A --watch=static/,routes/,islands/ www/dev.ts"
   },
   "importMap": "./tests/fixture/import_map.json"
 }
