+++ dev.ts
@@ -1,0 +1,2 @@
+import { dev } from "./src/dev/mod.ts";
+export default dev;
