+++ examples/counter/fresh.gen.ts
@@ -3,11 +3,13 @@
 // To update this file, run `fresh manifest`.
 
 import * as $0 from "./routes/index.tsx";
+import * as $1 from "./routes/index2.tsx";
 import * as $$0 from "./islands/Counter.tsx";
 
 const manifest = {
   routes: {
     "./routes/index.tsx": $0,
+    "./routes/index2.tsx": $1,
   },
   islands: {
     "./islands/Counter.tsx": $$0,
