--- src/cli/deps.ts
@@ -1,15 +1,0 @@
-// std
-export {
-  extname,
-  join,
-  resolve,
-  toFileUrl,
-} from "https://deno.land/std@0.128.0/path/mod.ts";
-export { walk } from "https://deno.land/std@0.128.0/fs/walk.ts";
-export { parse as parseArgs } from "https://deno.land/std@0.128.0/flags/mod.ts";
-export {
-  bold,
-  green,
-  red,
-  yellow,
-} from "https://deno.land/std@0.128.0/fmt/colors.ts";
