+++ src/dev/deps.ts
@@ -1,0 +1,11 @@
+// std
+export {
+  dirname,
+  extname,
+  fromFileUrl,
+  join,
+  resolve,
+  toFileUrl,
+} from "https://deno.land/std@0.128.0/path/mod.ts";
+export { walk } from "https://deno.land/std@0.128.0/fs/walk.ts";
+export { bold, red } from "https://deno.land/std@0.128.0/fmt/colors.ts";
