+++ www/deno.json
@@ -1,6 +1,6 @@
 {
   "tasks": {
-    "start": "deno run -A --watch main.ts & deno run -A ../cli.ts manifest --watch"
+    "start": "deno run -A --watch=static/,routes/,islands/ dev.ts"
   },
   "importMap": "./import_map.json"
 }
