+++ www/dev.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env -S deno run -A --watch=static/,routes/,islands/
+
+import dev from "$fresh/dev.ts";
+
+await dev(import.meta.url, "./main.ts");
